tests/krate/following: Move tests into `routes` submodules
