+++ src/tests/krate/following.rs
@@ -3,14 +3,6 @@ use crate::util::{RequestHelper, TestApp};
 use crate::OkBool;
 
 #[test]
-fn diesel_not_found_results_in_404() {
-    let (_, _, user) = TestApp::init().with_user();
-
-    user.get("/api/v1/crates/foo_following/following")
-        .assert_not_found();
-}
-
-#[test]
 fn following() {
     // TODO: Test anon requests as well?
     let (app, _, user) = TestApp::init().with_user();
@@ -56,23 +48,6 @@ fn following() {
     unfollow();
     assert!(!is_following());
     assert_eq!(user.search("following=1").crates.len(), 0);
-}
-
-#[test]
-fn disallow_api_token_auth_for_get_crate_following_status() {
-    let (app, _, _, token) = TestApp::init().with_token();
-    let api_token = token.as_model();
-
-    let a_crate = "a_crate";
-
-    app.db(|conn| {
-        CrateBuilder::new(a_crate, api_token.user_id).expect_build(conn);
-    });
-
-    // Token auth on GET for get following status is disallowed
-    token
-        .get(&format!("/api/v1/crates/{a_crate}/following"))
-        .assert_forbidden();
 }
 
 #[test]
