+++ src/tests/routes/crates/mod.rs
@@ -1,4 +1,5 @@
 pub mod downloads;
+mod following;
 mod new;
 mod read;
 pub mod versions;
